added save teacher
change data type long to int
save teacher api
get teacher by id api

+++ src/main/java/com/edunest/entity/ClassMaster.java
@@ -15,7 +15,7 @@ public class ClassMaster {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "class_id")
-    private Long classId;
+    private Integer classId;
 
     @Column(name = "tenant_id", nullable = false)
     private Integer tenantId;
