+++ src/main/java/com/edunest/entity/ClassSubject.java
@@ -15,19 +15,19 @@ public class ClassSubject {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "class_subject_id")
-    private Long classSubjectId;
+    private Integer classSubjectId;
 
     @Column(name = "tenant_id", nullable = false)
     private Integer tenantId;
 
     @Column(name = "class_id", nullable = false)
-    private Long classId;
+    private Integer classId;
 
     @Column(name = "subject_id", nullable = false)
-    private Long subjectId;
+    private Integer subjectId;
 
     @Column(name = "academic_year_id", nullable = false)
-    private Long academicYearId;
+    private Integer academicYearId;
 
     @Column(name = "is_active")
     @Builder.Default
