+++ src/main/java/com/edunest/entity/Subject.java
@@ -15,7 +15,7 @@ public class Subject {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "subject_id")
-    private Long subjectId;
+    private Integer subjectId;
 
     @Column(name = "tenant_id", nullable = false)
     private Integer tenantId;
